libm_dbl: several fixes for a DPFP-based build for the ATSAMD21.

+++ lib/libm_dbl/tgamma.c
@@ -116,7 +116,7 @@ double tgamma(double x)
 	/* special cases */
 	if (ix >= 0x7ff00000)
 		/* tgamma(nan)=nan, tgamma(inf)=inf, tgamma(-inf)=nan with invalid */
-		return x + INFINITY;
+		return x + (double)INFINITY;
 	if (ix < (0x3ff-54)<<20)
 		/* |x| < 2^-54: tgamma(x) ~ 1/x, +-0 raises div-by-zero */
 		return 1/x;
